Picking: Allow multiple picking at once
For cases like picking either a Tile or a Boat, the
current system wasn't valid as it would only allow
one type at a time.
This commit allows to pass both a TileRequest and an
EntityRequest to the picking request, removing this
limitation.
It is therefore now possible to move an entity into
a boat.